+++ packages/generator-uno-serverless/generators/app/templates/aws-lambda-sls/src/config.ts
@@ -1,0 +1,4 @@
+
+/** The configuration keys can be listed here. */
+export enum Config {
+}