+++ packages/generator-uno-serverless/generators/app/templates/aws-lambda-sls/src/handlers/health.ts
@@ -1,0 +1,10 @@
+import { Container, httpFunc } from "@common";
+import { health } from "uno-serverless";
+
+export const handler = httpFunc()
+  .handler(health<Container>(
+    "<%= projectName %>",
+    async ({ services }) => ([
+      services.configService() as any,
+    ]),
+  ));