+++ packages/generator-uno-serverless/generators/app/templates/azure/test/unit/handlers/common.test.ts
@@ -1,0 +1,12 @@
+import { createContainer } from "@common";
+import { expect } from "chai";
+
+describe("container", () => {
+
+  it("should instantiate container", async () => {
+    const container = createContainer();
+
+    expect(container.configService()).to.not.be.undefined;
+  });
+
+});